+++ lib/src/store.dart
@@ -1,7 +1,10 @@
 import 'dart:async';
+import 'package:rxdart/streams.dart';
 import 'package:rxdart/subjects.dart';
+import 'package:rxdart/transformers.dart';
 import 'package:stream_store/src/reducer.dart';
 import 'package:stream_store/src/states.dart';
+import 'package:stream_store/stream_store.dart';
 
 class Store<S> extends StreamView<S> implements Sink<dynamic> {
   final Sink<dynamic> _actions;
@@ -13,22 +16,46 @@ class Store<S> extends StreamView<S> implements Sink<dynamic> {
   factory Store(
     Reducer<S> reducer, {
     S initialState,
-    List<StreamTransformer<dynamic, dynamic>> transformers = const [],
+    List<Effect<S>> effects = const [],
   }) {
     // ignore: close_sinks
-    final actions = new StreamController.broadcast();
+    final actionsController = new StreamController.broadcast();
     // ignore: close_sinks
     final reducers = new BehaviorSubject<Reducer<S>>(seedValue: reducer);
     // ignore: close_sinks
-    final states = new States<S>(
-        initialState,
-        reducers.stream,
-        transformers.fold(
-          actions.stream,
-          (stream, transformer) => stream.transform(transformer),
-        ));
+    final states = new States<S>(initialState, reducers.stream);
 
-    return new Store._(actions, reducers, states);
+    final actions = effects.isEmpty
+        ? actionsController.stream
+        : actionsController.stream.transform(_buildEffects<S>(
+            effects,
+            states.stream,
+          ));
+
+    actions
+        .transform(
+          new WithLatestFromStreamTransformer(
+            reducers.stream,
+            (a, r) {
+              return new _ActionAndReducer<S>(a, r);
+            },
+          ),
+        )
+        .transform(
+          new ScanStreamTransformer(
+            (S state, _ActionAndReducer<S> latest, int i) {
+              return latest.reducer(state, latest.action);
+            },
+            initialState,
+          ),
+        )
+        .listen(
+          (val) => states.add(val),
+          onError: states.addError,
+          onDone: states.close,
+        );
+
+    return new Store._(actionsController, reducers, states);
   }
 
   void set reducer(Reducer<S> reducer) {
@@ -47,6 +74,48 @@ class Store<S> extends StreamView<S> implements Sink<dynamic> {
     _reducers.close();
     _states.close();
   }
+
+  static StreamTransformer<dynamic, dynamic> _buildEffects<S>(
+    List<Effect> effects,
+    Stream<S> states,
+  ) {
+    final combinedEffect = (states, actions) {
+      return new MergeStream(
+        // Note: `toList` is very important here. If you simply use `map`, the
+        // `MappedListIterable` can evaluate this function multiple times
+        // creating duplicated actions!
+        effects.map((effect) => effect(states, actions)).toList(),
+      );
+    };
+
+    return new StreamTransformer((Stream<dynamic> actions, bool cancelOnError) {
+      final controller = new StreamController<dynamic>.broadcast();
+
+      actions.listen(
+        (item) {
+          controller.add(item);
+        },
+        onError: controller.addError,
+      );
+
+      combinedEffect(states, actions).listen(
+        (item) {
+          controller.add(item);
+        },
+        onError: controller.addError,
+        onDone: controller.close,
+      );
+
+      return controller.stream.listen(null);
+    });
+  }
 }
 
 enum StreamStoreActions { REPLACE }
+
+class _ActionAndReducer<S> {
+  final dynamic action;
+  final Reducer<S> reducer;
+
+  _ActionAndReducer(this.action, this.reducer);
+}