+++ lib/stream_store.dart
@@ -3,4 +3,4 @@ library stream_store;
 export 'src/reducer.dart';
 export 'src/effect.dart';
 export 'src/store.dart';
-export 'src/logging_transformer.dart';
+export 'src/logging_effect.dart';