+++ test/src/effect_test.dart
@@ -4,11 +4,14 @@ import "test_utils.dart";
 
 void main() {
   group("Effects", () {
-    Object identityReducer(Object state, Object action) => action;
+    TestActions identityReducer(TestActions state, dynamic action) =>
+        action is TestActions ? action : state;
 
     test("transform but cannot swallow actions", () async {
-      final transformer = new EffectTransformer(send1Effect);
-      final store = new Store(identityReducer, transformers: [transformer]);
+      final store = new Store<TestActions>(
+        identityReducer,
+        effects: [send1Effect],
+      );
 
       store.add(TestActions.SEND1);
 
@@ -17,11 +20,9 @@ void main() {
     });
 
     test("can be combined", () async {
-      final store = new Store(identityReducer, transformers: [
-        new EffectTransformer.combine([
-          send1Effect,
-          send2Effect,
-        ]),
+      final store = new Store(identityReducer, effects: [
+        send1Effect,
+        send2Effect,
       ]);
 
       store.add(TestActions.SEND1);
@@ -37,25 +38,10 @@ void main() {
           ]));
     });
 
-    test("more than one transformer is supported", () async {
-      final transformer1 = new EffectTransformer(send1Effect);
-      final transformer2 = new EffectTransformer(send2Effect);
-      final store = new Store(identityReducer, transformers: [
-        transformer1,
-        transformer2,
-      ]);
-
-      store.add(TestActions.SEND1);
-
-      await expect(
-          store, emitsInOrder([TestActions.SEND1, TestActions.RESPOND1]));
-    });
-
     test("work with async streams", () async {
-      final transformer = new EffectTransformer(cancellableResponse);
       final store = new Store(
         identityReducer,
-        transformers: [transformer],
+        effects: [cancellableResponse],
       );
 
       store.add(TestActions.SEND1);
@@ -72,10 +58,9 @@ void main() {
     });
 
     test("can be cancelled by dispatching follow up actions", () async {
-      final transformer = new EffectTransformer(cancellableResponse);
       final store = new Store(
         identityReducer,
-        transformers: [transformer],
+        effects: [cancellableResponse],
       );
 
       store.add(TestActions.SEND1);
@@ -93,10 +78,9 @@ void main() {
     });
 
     test("can send multiple actions in response to a single action", () async {
-      final transformer = new EffectTransformer(respondTwiceEffect);
       final store = new Store(
         identityReducer,
-        transformers: [transformer],
+        effects: [respondTwiceEffect],
       );
 
       store.add(TestActions.SEND1);
@@ -108,6 +92,27 @@ void main() {
             TestActions.SEND1,
             TestActions.RESPOND1,
             TestActions.RESPOND2,
+          ],
+        ),
+      );
+    });
+
+    test("can read the state", () async {
+      final store = new Store(
+        identityReducer,
+        initialState: TestActions.RESPOND1,
+        effects: [readStateEffect],
+      );
+
+      store.add(TestActions.SEND1);
+
+      await expect(
+        store,
+        emitsInOrder(
+          [
+            TestActions.RESPOND1,
+            TestActions.SEND1,
+            TestActions.RESPOND1,
           ],
         ),
       );